Made more components from the multiform component, making the code slightly more readable
